requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-rest-knox==4.1.0
django-summernote==0.8.11.6
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
Pillow==8.0.1
prompt-toolkit==3.0.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
